depart verifier code
